requirements.txt
Pillow
tgtg
requests
tkcalendar
opencv-python
pyzbar
numpy 
